feat: create resetPassword/sendEmail pages/context, refactor address
